+++ internal/output/error_test.go
@@ -23,6 +23,27 @@ func TestError(t *testing.T) {
 			expectedJSONResult:  "{\n  \"error\": \"MOMO\"\n}",
 			expectedYAMLResult:  "error: MOMO\n",
 		},
+		{
+			name:                "error with resource",
+			input:               output.Error{Value: fmt.Errorf("MOMO"), Resolved: "resourceName"},
+			expectedHumanResult: "\nERROR: MOMO",
+			expectedJSONResult:  "{\n  \"error\": \"MOMO\",\n  \"resource\": \"resourceName\"\n}",
+			expectedYAMLResult:  "error: MOMO\nresource: resourceName\n",
+		},
+		{
+			name:                "error with argument",
+			input:               output.Error{Value: fmt.Errorf("MOMO"), Original: "testArg"},
+			expectedHumanResult: "\nERROR: MOMO (testArg)",
+			expectedJSONResult:  "{\n  \"argument\": \"testArg\",\n  \"error\": \"MOMO\"\n}",
+			expectedYAMLResult:  "argument: testArg\nerror: MOMO\n",
+		},
+		{
+			name:                "error with argument and resource",
+			input:               output.Error{Value: fmt.Errorf("MOMO"), Resolved: "resourceName", Original: "testArg"},
+			expectedHumanResult: "\nERROR: MOMO (testArg)",
+			expectedJSONResult:  "{\n  \"argument\": \"testArg\",\n  \"error\": \"MOMO\",\n  \"resource\": \"resourceName\"\n}",
+			expectedYAMLResult:  "argument: testArg\nerror: MOMO\nresource: resourceName\n",
+		},
 	}
 	for _, test := range errorTests {
 		t.Run(test.name, test.Generate())